Got rid of one hardcoded value
