--- include/functions.h
@@ -1,11 +1,0 @@
-#ifndef FUNCTIONS_H
-#define FUNCTIONS_H
-
-
-#include <SDL3/SDL_pixels.h>
-#include <SDL3/SDL_render.h>
- 
-unsigned char generateRandomNN(int mask);
-
-void cleanup();
-#endif
