+++ makefile
@@ -4,7 +4,7 @@ LDFLAGS = $(shell pkg-config --libs sdl3 sdl3-ttf)
 
 
 SRC_DIR = src
-SRC = $(SRC_DIR)/main.c $(SRC_DIR)/chip8.c $(SRC_DIR)/functions.c $(SRC_DIR)/config.c $(SRC_DIR)/overlay.c $(SRC_DIR)/audio.c
+SRC = $(SRC_DIR)/main.c $(SRC_DIR)/chip8.c $(SRC_DIR)/config.c $(SRC_DIR)/overlay.c $(SRC_DIR)/audio.c
 OBJ_DIR = obj
 OBJ = $(SRC:$(SRC_DIR)/%.c=$(OBJ_DIR)/%.o)
