--- src/functions.c
@@ -1,30 +1,0 @@
-#include <SDL3_ttf/SDL_ttf.h>
-#include <stdlib.h>
-
-#include <SDL3/SDL.h>
-#include <SDL3/SDL_timer.h>
-#include <SDL3/SDL_events.h>
-#include <SDL3/SDL_init.h>
-#include <SDL3/SDL_keyboard.h>
-#include <SDL3/SDL_render.h>
-#include <SDL3/SDL_scancode.h>
-#include <SDL3/SDL_log.h>
-
-#include "../include/functions.h"
-#include "../include/config.h"
-
-extern uint8_t memory[4096]; //the total memory of the chip-8
-extern Config *globalConfig;
-
-
-unsigned char generateRandomNN(int mask){
-    int randomNumber = rand() % (255 + 1 - 0) + 0;
-    return randomNumber & mask;
-}
-
-void cleanup(){
-    SDL_LogInfo(SDL_LOG_CATEGORY_APPLICATION, "Quitting emul8...\n");
-    TTF_Quit();
-    SDL_Quit();
-    exit(0);
-}
